+++ packages/steps/src/components/step-indicator/Component.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import React from 'react';
 import cn from 'classnames';
 import { Badge, BadgeProps } from '@alfalab/core-components-badge';
 
@@ -7,13 +7,13 @@ import styles from './index.module.css';
 export type StepIndicatorProps = Pick<BadgeProps, 'content' | 'iconColor' | 'className'>;
 
 export const StepIndicator: React.FC<StepIndicatorProps> = ({ content, iconColor, className }) => {
-  return (
-    <Badge
-        size='l'
-        view='icon'
-        iconColor={iconColor}
-        className={cn(styles.component, className)}
-        content={content}
-    />
-  )
-}
+    return (
+        <Badge
+            size='l'
+            view='icon'
+            iconColor={iconColor}
+            className={cn(styles.component, className)}
+            content={content}
+        />
+    );
+};