+++ CHANGELOG.md
@@ -1,3 +1,48 @@
+## 32.1.0
+
+### Minor Changes
+
+#### `@alfalab/core-components-date-range-input`
+- [#175](https://github.com/core-ds/core-components/pull/175): Новые компоненты: TimeInput, DateTimeInput, DateRangeInput. Thanks [@blackraydev](https://github.com/blackraydev)
+
+
+#### `@alfalab/core-components-date-time-input`
+- [#175](https://github.com/core-ds/core-components/pull/175): Новые компоненты: TimeInput, DateTimeInput, DateRangeInput. Thanks [@blackraydev](https://github.com/blackraydev)
+
+
+#### `@alfalab/core-components-time-input`
+- [#175](https://github.com/core-ds/core-components/pull/175): Новые компоненты: TimeInput, DateTimeInput, DateRangeInput. Thanks [@blackraydev](https://github.com/blackraydev)
+
+
+#### `@alfalab/core-components-themes`
+- [#269](https://github.com/core-ds/core-components/pull/269): Удалена темизация click для компонента CodeInput. Thanks [@Valeri8888](https://github.com/Valeri8888)
+
+
+### Patch Changes
+
+#### `@alfalab/core-components-calendar`
+- [#175](https://github.com/core-ds/core-components/pull/175): Новые компоненты: TimeInput, DateTimeInput, DateRangeInput. Thanks [@blackraydev](https://github.com/blackraydev)
+
+
+#### `@alfalab/core-components-code-input`
+- [#274](https://github.com/core-ds/core-components/pull/274): Компактная версия теперь включается когда >6 полей. Thanks [@Valeri8888](https://github.com/Valeri8888)
+
+## 32.0.1
+
+### Patch Changes
+
+#### `@alfalab/core-components-pure-cell`
+- [#271](https://github.com/core-ds/core-components/pull/271): Внутренний компонент "comment" заменен на @alfalab/core-components-comment. Thanks [@Lacronts](https://github.com/Lacronts)
+- [#272](https://github.com/core-ds/core-components/pull/272): HTML-тег main заменен на div. Thanks [@Lacronts](https://github.com/Lacronts)
+
+
+#### `@alfalab/core-components-side-panel`
+- [#270](https://github.com/core-ds/core-components/pull/270): Исправлена ошибка, когда non-static position контент при скролле перекрывал sticky заголовок. Thanks [@Lacronts](https://github.com/Lacronts)
+
+
+#### `@alfalab/core-components-vars`
+- [#273](https://github.com/core-ds/core-components/pull/273): Исправлены пути импорта переменных в бандле. Thanks [@Lacronts](https://github.com/Lacronts)
+
 ## 32.0.0
 
 ### Major Changes