+++ packages/calendar-input/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 7.1.4
+
+### Patch Changes
+
+-   Updated dependencies [[#175](https://github.com/core-ds/core-components/pull/175)]
+    -   @alfalab/core-components-calendar@6.1.4
+
 ## 7.1.3
 
 ### Patch Changes